задача4 часть1
